Update connector to accept multidimensional arrays OR files
The current connector, by never calling http_build_query, could no longer
accept multidimensional arrays - which are requested by most of our
endpoints. However, http_build_query does not work with cURL file
submissions (which is why it was removed).
Update the connector to use the raw array when files are present
(which will discard any dimensions beyond the first in the submission),
or to build the query otherwise.